requirement.txt
streamlit
pandas
numpy
matplotlib
requests
python-dotenv
pdfplumber
chardet
plotly 